requirements.txt
selenium==4.27.1
groq==0.15.0
python-telegram-bot==21.9
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
